requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-Login==0.5.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
python-dotenv==0.19.0
requests==2.26.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
